feat(users): implement GET all users at /api/users
