+++ database/config/config.json
@@ -1,13 +1,18 @@
-
 {
   "development": {
-    "username": "postgres" ,
+    "username": "postgres",
     "password": null,
-
+    "database": "postgres",
+    "host": "127.0.0.1",
+    "dialect": "postgres",
+    "port": 5433
+  },
+  "test": {
+    "username": "postgres",
+    "password": null,
     "database": "postgres",
     "host": "127.0.0.1",
     "dialect": "postgres",
     "port": 5433
   }
 }
-
