+++ src/models/users.js
@@ -29,11 +29,12 @@ module.exports = (sequelize, DataTypes) => {
     skills: DataTypes.ARRAY(DataTypes.STRING),
     role: DataTypes.ENUM('unspecified'),
     guild: DataTypes.ENUM('unspecified'),
-    past_engagement_id: DataTypes.ARRAY(DataTypes.STRING),
+    past_engagement_ids: DataTypes.ARRAY(DataTypes.STRING),
     image: DataTypes.STRING
   }, {
     sequelize,
     modelName: 'users',
+    underscored: true,
   });
   return Users;
 };
