+++ src/routes/user.routes.js
@@ -1,0 +1,10 @@
+
+const express = require('express');
+const { getUsers } = require('../controllers/user.controller');
+
+const router = express.Router();
+
+router.get('/users', getUsers);
+
+
+module.exports = router;
